Fixed add-bookmark-bug.
When adding a bookmark using the 'Add Bookmark'-
button, there was a bug where
the DocumentArea would detect the
MouseEvent.MOUSE_CLICKED as a separate click
after adding a bookmark. This resulted in
clicking outside of the address bar which
causes the URL in the AddressBar to be loaded
uncorrectly. Thanks Martijn ;)